+++ include/util/StringSampleIter.hpp
@@ -59,7 +59,7 @@ namespace vmm_util {
         }
 
 
-        long size() {
+        size_t size() {
             return data.size();
         }
